# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Statera API (Partial Static Spec)
  version: 1.0.0
  description: >-
    Partial static OpenAPI contract for the Flask backend. It does not yet cover
    health probes, worker-health endpoints, feature flags, messages,
    notifications, categories, merchants, budgets, memorized transactions, or
    the open-banking route group.
  license:
    name: Proprietary
servers:
  - url: /
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  parameters:
...
    goalId:
      name: goal_id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
...
    ValidationError:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    GenericObject:
      type: object
      additionalProperties: true
    OkEnvelope:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/GenericObject'
        meta:
          $ref: '#/components/schemas/GenericObject'
      additionalProperties: true
      required:
        - ok
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
        errors:
          type: array
          items:
            type: string
        code:
          type: string
        error_code:
          type: string
        retry_after:
          type: integer
      additionalProperties: true
...
    SavingsGoal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        goal_type:
          type: string
        target_kd:
          type: string
        current_kd:
          type: string
        target_date:
          type: string
          format: date
          nullable: true
        linked_category:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        is_active:
          type: boolean
        projection:
          $ref: '#/components/schemas/GenericObject'
      additionalProperties: true
    SavingsGoalRequest:
      type: object
      properties:
        name:
          type: string
        goal_type:
          type: string
        target_kd:
          type: string
        current_kd:
          type: string
        target_date:
          type: string
          format: date
          nullable: true
        linked_category:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      additionalProperties: true
paths:
...
  /api/savings-goals:
    get:
      tags: [Goals]
      summary: List savings goals.
      operationId: listSavingsGoals
      security:
        - sessionCookie: []
      parameters:
        - name: include_inactive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Savings goals list.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OkEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          goals:
                            type: array
                            items:
                              $ref: '#/components/schemas/SavingsGoal'
                          include_inactive:
                            type: boolean
                        additionalProperties: true
    post:
      tags: [Goals]
      summary: Create a savings goal.
      operationId: createSavingsGoal
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavingsGoalRequest'
      responses:
        '201':
          description: Savings goal created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/savings-goals/{goal_id}/projection:
    get:
      tags: [Goals]
      summary: Return projection details for a savings goal.
      operationId: getSavingsGoalProjection
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/goalId'
      responses:
        '200':
          description: Projection payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/savings-goals/{goal_id}/update:
    post:
      tags: [Goals]
      summary: Update a savings goal.
      operationId: updateSavingsGoal
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/goalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavingsGoalRequest'
      responses:
        '200':
          description: Savings goal updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/savings-goals/{goal_id}/deposit:
    post:
      tags: [Goals]
      summary: Deposit progress into a savings goal.
      operationId: depositSavingsGoal
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/goalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount_kd
              properties:
                amount_kd:
                  type: string
      responses:
        '200':
          description: Savings goal progress updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/savings-goals/{goal_id}/delete:
    post:
      tags: [Goals]
      summary: Soft-delete a savings goal.
      operationId: deleteSavingsGoal
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/goalId'
      responses:
        '200':
          description: Savings goal archived.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'